Address precision=16 warning
"python3.11/site-packages/lightning/fabric/connector.py:571: `precision=16` is supported for historical reasons but its usage is discouraged. Please set your precision to 16-mixed instead!""

+++ src/python/epiclass/epiatlas_training_no_valid.py
@@ -277,7 +277,7 @@ def train_without_valid(
                 callbacks=callbacks,
                 accelerator="gpu",
                 devices=1,
-                precision=16,
+                precision="16-mixed",
             )
         else:
             trainer = MyTrainer(
